New RAMSES release, Dec. 2018
The latest version of RAMSES, delivered as singular patch.

New features (refer to documentation for more info):
- scene expiration tracking
- Fine-grained control over rendering interruption limits

API changes:
- removed Translate, Scane etc. nodes - there is just Node which has all
    properties

Improvements:
- Local scenes are never force-unsubscribed
- Improved TCP communication stack
- Using more C++11
- Better performance of Text API
- Improved logging
- Simplified directory structure of client and renderer module
- various CMake improvements
- Removed obsolete code
- Improved code style checker scripts

Original internal commit sha:
c02a0c639e266bc9fbb8a0871ae029d31df72f52

+++ client/ramses-client/impl/ArrayResourceImpl.cpp
@@ -32,7 +32,7 @@ namespace ramses
         CHECK_RETURN_ERR(ResourceImpl::serialize(outStream, serializationContext));
 
         outStream << m_elementCount;
-        outStream << m_elementType;
+        outStream << static_cast<uint32_t>(m_elementType);
 
         return StatusOK;
     }
